fix(mobile): polish discovery and home interactions
