+++ apps/mobile/src/features/home/home.styles.ts
@@ -101,6 +101,11 @@ export const styles = StyleSheet.create({
     shadowOffset: { width: spacing.none, height: spacing.xs },
     elevation: 12,
   },
+  unifiedRailCreateButtonStroke: {
+    ...StyleSheet.absoluteFillObject,
+    borderRadius: radius['2xl'],
+    borderWidth: StyleSheet.hairlineWidth,
+  },
   unifiedRailServerTouch: {
     width: size.plusPanelIcon + spacing.sm,
     height: size.plusPanelIcon + spacing.sm,
@@ -108,14 +113,23 @@ export const styles = StyleSheet.create({
     justifyContent: 'center',
     overflow: 'visible',
   },
+  unifiedRailAvatarShell: {
+    width: size.plusPanelIcon,
+    height: size.plusPanelIcon,
+    borderRadius: radius['2xl'],
+    alignItems: 'center',
+    justifyContent: 'center',
+    overflow: 'visible',
+  },
   unifiedRailAvatar: {
     width: size.plusPanelIcon,
     height: size.plusPanelIcon,
-    borderRadius: radius.xl,
-    borderWidth: border.active,
-    alignItems: 'center',
-    justifyContent: 'center',
+    borderRadius: radius['2xl'],
     overflow: 'hidden',
+  },
+  unifiedRailAvatarOutline: {
+    ...StyleSheet.absoluteFillObject,
+    borderRadius: radius['2xl'],
   },
   unifiedRailDirectAvatar: {
     width: size.plusPanelIcon,
@@ -135,6 +149,17 @@ export const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'center',
     paddingHorizontal: spacing.xs,
+  },
+  unifiedRailPrivateBadge: {
+    position: 'absolute',
+    top: spacing.none,
+    right: spacing.none,
+    width: size.badgeMd,
+    height: size.badgeMd,
+    borderRadius: radius.full,
+    borderWidth: StyleSheet.hairlineWidth,
+    alignItems: 'center',
+    justifyContent: 'center',
   },
   unifiedRailBadgeText: {
     color: palette.white,
@@ -654,6 +679,21 @@ export const styles = StyleSheet.create({
     width: size.avatarMd,
     height: size.avatarMd,
     borderRadius: radius.lg,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  commandModalAvatarWrap: {
+    width: size.avatarMd,
+    height: size.avatarMd,
+  },
+  commandModalPrivateBadge: {
+    position: 'absolute',
+    top: -spacing.xxs,
+    right: -spacing.xxs,
+    width: size.badgeSm,
+    height: size.badgeSm,
+    borderRadius: radius.full,
+    borderWidth: StyleSheet.hairlineWidth,
     alignItems: 'center',
     justifyContent: 'center',
   },
